requirements.txt
yt-dlp
moviepy>=2.0
librosa
numpy
openai-whisper
schedule
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
instagrapi
